api/requirements.txt
black==19.3b0
channels==2.3.0
channels-redis==2.4.0
daphne==2.3.0
django-allauth==0.40.0
django-debug-toolbar==2.0
django-filter==2.2.0
django-guardian==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
factory-boy==2.12.0
flake8==3.7.8
psycopg2-binary==2.8.4
texttable==1.6.2
